fix: changed rank name, removed unstable features

+++ MemberSystem/src/main/java/com/isacariotsystems/MemberSystem/repository/MemberRepository.java
@@ -11,8 +11,9 @@ import com.isacariotsystems.MemberSystem.entity.Member;
 @Repository
 public interface MemberRepository extends JpaRepository<Member,Long>{
 
+    /* 
     List<Member> findByBranchId(Long branchId);
 
-    List<Member> findByRankId(Integer rankId);
-    
+    List<Member> findByRankId(Long rankId);
+    */
 }
