+++ MemberSystem/src/main/java/com/isacariotsystems/MemberSystem/service/RankService.java
@@ -11,9 +11,9 @@ public interface RankService {
 
     public List<Rank> allRanks();
 
-    public Optional<Rank> findRankById(Integer rankId);
+    public Optional<Rank> findRankById(Long rankId);
 
-    public void deleteRankById(Integer rankId);
+    public void deleteRankById(Long rankId);
 
-    public Rank updateRank(Integer rankId, Rank rank);
+    public Rank updateRank(Long rankId, Rank rank);
 }
